+++ packages/orq-rc/RELEASES.md
@@ -429,3 +429,13 @@ Based on:
 - [typescript v3.1.0-rc.63] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.63] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.63 - packages/orq-rc
+
+## 2025-02-11 15:59:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.64] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.64 - packages/orq-rc